Rename recordset.hpp as the content is actually quite low level.
