+++ tests/compile/include/record_block.cpp
@@ -3,5 +3,5 @@
 */
 
 
-#include <pgasio/recordset.hpp>
+#include <pgasio/record_block.hpp>
